Pruebas de interfaz añadidas al respectivo namespace. Nombres de archivos cambiados para que representen sus contenidos.
