--- CI0126-ExamenFinal-B70866.Tests/UITesting/UnitTest2.cs
@@ -1,32 +1,0 @@
-﻿// Generated by Selenium IDE
-using System.Collections.Generic;
-using OpenQA.Selenium;
-using OpenQA.Selenium.Chrome;
-using NUnit.Framework;
-[TestFixture]
-public class CPAddToCartTest
-{
-    private IWebDriver driver;
-    public IDictionary<string, object> vars { get; private set; }
-    private IJavaScriptExecutor js;
-    [SetUp]
-    public void SetUp()
-    {
-        driver = new ChromeDriver();
-        js = (IJavaScriptExecutor)driver;
-        vars = new Dictionary<string, object>();
-    }
-    [TearDown]
-    protected void TearDown()
-    {
-        driver.Quit();
-    }
-    [Test]
-    public void cPAddToCart()
-    {
-        driver.Navigate().GoToUrl("https://localhost:44338/");
-        driver.Manage().Window.Size = new System.Drawing.Size(1296, 829);
-        driver.FindElement(By.LinkText("Ver Productos")).Click();
-        driver.FindElement(By.LinkText("Comprar individualmente")).Click();
-    }
-}
